#2144 - Project state change events only work for matrix workload
- Move project state watcher to matrix workflow module
- Merge monitoring ui module into matrix workflow module
