+++ inception/inception-app-webapp/pom.xml
@@ -291,10 +291,6 @@
     </dependency>
     <dependency>
       <groupId>de.tudarmstadt.ukp.inception.app</groupId>
-      <artifactId>inception-ui-monitoring</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>de.tudarmstadt.ukp.inception.app</groupId>
       <artifactId>inception-remote</artifactId>
     </dependency>
     <dependency>
