+++ inception/inception-project/src/main/java/de/tudarmstadt/ukp/clarin/webanno/project/ProjectServiceImpl.java
@@ -17,11 +17,6 @@
  */
 package de.tudarmstadt.ukp.clarin.webanno.project;
 
-import static de.tudarmstadt.ukp.clarin.webanno.model.SourceDocumentState.ANNOTATION_FINISHED;
-import static de.tudarmstadt.ukp.clarin.webanno.model.SourceDocumentState.ANNOTATION_IN_PROGRESS;
-import static de.tudarmstadt.ukp.clarin.webanno.model.SourceDocumentState.CURATION_FINISHED;
-import static de.tudarmstadt.ukp.clarin.webanno.model.SourceDocumentState.CURATION_IN_PROGRESS;
-import static de.tudarmstadt.ukp.clarin.webanno.model.SourceDocumentState.NEW;
 import static java.nio.file.Files.newDirectoryStream;
 import static java.util.Arrays.asList;
 import static java.util.stream.Collectors.toList;
@@ -41,7 +36,6 @@ import java.nio.file.DirectoryStream;
 import java.nio.file.NoSuchFileException;
 import java.nio.file.Path;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.Date;
@@ -50,14 +44,12 @@ import java.util.HashSet;
 import java.util.LinkedHashSet;
 import java.util.LinkedList;
 import java.util.List;
-import java.util.Objects;
 import java.util.Set;
 import java.util.stream.Collectors;
 
 import javax.persistence.EntityManager;
 import javax.persistence.NoResultException;
 import javax.persistence.PersistenceContext;
-import javax.persistence.Query;
 
 import org.apache.commons.collections4.SetUtils;
 import org.apache.commons.io.FileUtils;
@@ -75,21 +67,17 @@ import org.springframework.context.event.ContextRefreshedEvent;
 import org.springframework.context.event.EventListener;
 import org.springframework.core.annotation.AnnotationAwareOrderComparator;
 import org.springframework.stereotype.Component;
-import org.springframework.transaction.annotation.Isolation;
 import org.springframework.transaction.annotation.Transactional;
 
 import de.tudarmstadt.ukp.clarin.webanno.api.ProjectService;
 import de.tudarmstadt.ukp.clarin.webanno.api.RepositoryProperties;
 import de.tudarmstadt.ukp.clarin.webanno.api.event.AfterProjectCreatedEvent;
 import de.tudarmstadt.ukp.clarin.webanno.api.event.BeforeProjectRemovedEvent;
-import de.tudarmstadt.ukp.clarin.webanno.api.event.ProjectStateChangedEvent;
 import de.tudarmstadt.ukp.clarin.webanno.api.project.ProjectInitializer;
 import de.tudarmstadt.ukp.clarin.webanno.model.AnnotationDocumentState;
 import de.tudarmstadt.ukp.clarin.webanno.model.PermissionLevel;
 import de.tudarmstadt.ukp.clarin.webanno.model.Project;
 import de.tudarmstadt.ukp.clarin.webanno.model.ProjectPermission;
-import de.tudarmstadt.ukp.clarin.webanno.model.ProjectState;
-import de.tudarmstadt.ukp.clarin.webanno.model.SourceDocument;
 import de.tudarmstadt.ukp.clarin.webanno.security.UserDao;
 import de.tudarmstadt.ukp.clarin.webanno.security.model.Authority;
 import de.tudarmstadt.ukp.clarin.webanno.security.model.User;
@@ -163,98 +151,6 @@ public class ProjectServiceImpl
     public void updateProject(Project aProject)
     {
         entityManager.merge(aProject);
-    }
-
-    @Override
-    @Transactional(isolation = Isolation.READ_UNCOMMITTED)
-    public void recalculateProjectState(Project aProject)
-    {
-        Project project;
-        try {
-            project = getProject(aProject.getId());
-        }
-        catch (NoResultException e) {
-            // This happens when this method is called as part of deleting an entire project.
-            // In such a case, the project may no longer be available, so there is no point in
-            // updating its state. So then we do nothing here.
-            return;
-        }
-
-        // This query is better because we do not inject strings into the query string, but it
-        // does not work on HSQLDB (on MySQL it seems to work).
-        // See: https://github.com/webanno/webanno/issues/1011
-        // String query =
-        // "SELECT new " + SourceDocumentStateStats.class.getName() + "(" +
-        // "COUNT(*) AS num, " +
-        // "SUM(CASE WHEN state = :an THEN 1 ELSE 0 END), " +
-        // "SUM(CASE WHEN (state = :aip OR state is NULL) THEN 1 ELSE 0 END), " +
-        // "SUM(CASE WHEN state = :af THEN 1 ELSE 0 END), " +
-        // "SUM(CASE WHEN state = :cip THEN 1 ELSE 0 END), " +
-        // "SUM(CASE WHEN state = :cf THEN 1 ELSE 0 END)) " +
-        // "FROM SourceDocument " +
-        // "WHERE project = :project";
-        //
-        // SourceDocumentStateStats stats = entityManager.createQuery(
-        // query, SourceDocumentStateStats.class)
-        // .setParameter("project", aProject)
-        // .setParameter("an", SourceDocumentState.NEW)
-        // .setParameter("aip", SourceDocumentState.ANNOTATION_IN_PROGRESS)
-        // .setParameter("af", SourceDocumentState.ANNOTATION_FINISHED)
-        // .setParameter("cip", SourceDocumentState.CURATION_IN_PROGRESS)
-        // .setParameter("cf", SourceDocumentState.CURATION_FINISHED)
-        // .getSingleResult();
-
-        // @formatter:off
-        String query = 
-                "SELECT new " + SourceDocumentStateStats.class.getName() + "(" +
-                "COUNT(*), " +
-                "SUM(CASE WHEN state = '" + NEW.getId() + "'  THEN 1 ELSE 0 END), " +
-                "SUM(CASE WHEN (state = '" + ANNOTATION_IN_PROGRESS.getId() + 
-                        "' OR state is NULL) THEN 1 ELSE 0 END), " +
-                "SUM(CASE WHEN state = '" + ANNOTATION_FINISHED.getId() + 
-                        "'  THEN 1 ELSE 0 END), " +
-                "SUM(CASE WHEN state = '" + CURATION_IN_PROGRESS.getId() + 
-                        "' THEN 1 ELSE 0 END), " +
-                "SUM(CASE WHEN state = '" + CURATION_FINISHED.getId() + "'  THEN 1 ELSE 0 END)) " +
-                "FROM SourceDocument " + 
-                "WHERE project = :project";
-        // @formatter:on
-
-        SourceDocumentStateStats stats = entityManager
-                .createQuery(query, SourceDocumentStateStats.class)
-                .setParameter("project", aProject).getSingleResult();
-
-        ProjectState oldState = project.getState();
-
-        // We had some strange reports about being unable to calculate the project state, so to
-        // be better able to debug this, we add some more detailed information to the exception
-        // message here.
-        try {
-            project.setState(stats.getProjectState());
-        }
-        catch (IllegalStateException e) {
-            StringBuilder sb = new StringBuilder();
-            sb.append("\nDetailed document states in project [" + aProject.getName() + "]("
-                    + aProject.getId() + "):\n");
-            String detailQuery = "SELECT id, name, state FROM " + SourceDocument.class.getName()
-                    + " WHERE project = :project";
-            Query q = entityManager.createQuery(detailQuery).setParameter("project", aProject);
-            for (Object res : q.getResultList()) {
-                sb.append("- ");
-                sb.append(Arrays.toString((Object[]) res));
-                sb.append('\n');
-            }
-            IllegalStateException ne = new IllegalStateException(e.getMessage() + sb, e.getCause());
-            ne.setStackTrace(e.getStackTrace());
-            throw ne;
-        }
-
-        if (!Objects.equals(oldState, project.getState())) {
-            applicationEventPublisher
-                    .publishEvent(new ProjectStateChangedEvent(this, project, oldState));
-        }
-
-        updateProject(project);
     }
 
     @Override
