+++ inception/inception-workload-matrix/src/main/java/de/tudarmstadt/ukp/inception/workload/matrix/management/support/SourceDocumentActionsColumn.java
@@ -15,45 +15,26 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.tudarmstadt.ukp.clarin.webanno.ui.monitoring.support;
-
-import static org.apache.wicket.event.Broadcast.BUBBLE;
+package de.tudarmstadt.ukp.inception.workload.matrix.management.support;
 
 import org.apache.wicket.MarkupContainer;
 import org.apache.wicket.core.request.handler.IPageRequestHandler;
 import org.apache.wicket.extensions.markup.html.repeater.data.grid.ICellPopulator;
 import org.apache.wicket.extensions.markup.html.repeater.data.table.LambdaColumn;
-import org.apache.wicket.markup.html.form.CheckBox;
 import org.apache.wicket.markup.html.panel.Fragment;
 import org.apache.wicket.markup.repeater.Item;
-import org.apache.wicket.model.CompoundPropertyModel;
 import org.apache.wicket.model.IModel;
 import org.apache.wicket.model.Model;
 import org.apache.wicket.request.cycle.RequestCycle;
 
-import de.tudarmstadt.ukp.clarin.webanno.support.lambda.LambdaAjaxFormComponentUpdatingBehavior;
-import de.tudarmstadt.ukp.clarin.webanno.ui.monitoring.event.DocumentRowSelectionChangedEvent;
-
-public class SourceDocumentSelectColumn
-    extends LambdaColumn<DocumentMatrixRow, DocumentMatrixSortKey>
+public class SourceDocumentActionsColumn
+    extends LambdaColumn<DocumentMatrixRow, Void>
 {
     private static final long serialVersionUID = 8324173231787296215L;
 
-    private boolean visible;
-
-    public SourceDocumentSelectColumn()
+    public SourceDocumentActionsColumn()
     {
-        super(Model.of(""), row -> row.getSourceDocument().getName());
-    }
-
-    public void setVisible(boolean aVisible)
-    {
-        visible = aVisible;
-    }
-
-    public boolean isVisible()
-    {
-        return visible;
+        super(Model.of("Document"), row -> row.getSourceDocument().getName());
     }
 
     @Override
@@ -62,13 +43,6 @@ public class SourceDocumentSelectColumn
     {
         MarkupContainer page = IPageRequestHandler
                 .getPage(RequestCycle.get().getActiveRequestHandler());
-        Fragment frag = new Fragment(aComponentId, "select-column", page, aRowModel);
-        IModel<Boolean> selectedModel = CompoundPropertyModel.of(aRowModel).bind("selected");
-        CheckBox checkbox = new CheckBox("selected", selectedModel);
-        checkbox.add(new LambdaAjaxFormComponentUpdatingBehavior("change", _target -> checkbox
-                .send(checkbox, BUBBLE, new DocumentRowSelectionChangedEvent(_target, aRowModel))));
-        frag.add(checkbox);
-        frag.setVisible(visible);
-        aItem.add(frag);
+        aItem.add(new Fragment(aComponentId, "actions-column", page, aRowModel));
     }
 }
