+++ inception/inception-workload-matrix/src/main/resources/META-INF/asciidoc/user-guide/images/monitoring-annotation-states.png
