+++ inception/pom.xml
@@ -340,11 +340,6 @@
       </dependency>
       <dependency>
         <groupId>de.tudarmstadt.ukp.inception.app</groupId>
-        <artifactId>inception-ui-monitoring</artifactId>
-        <version>0.20.0-SNAPSHOT</version>
-      </dependency>
-      <dependency>
-        <groupId>de.tudarmstadt.ukp.inception.app</groupId>
         <artifactId>inception-ui-project</artifactId>
         <version>0.20.0-SNAPSHOT</version>
       </dependency>
